Build Main 70%완료
1. 사진 추가
2. 실데이터 삽입

todo
1. recycle view 간격 추가
2. text 글자 넘을시, 띄어쓰기
3. 원형이미지 background, 뒤에 색깔 입히기
4.
