+++ app/build.gradle
@@ -38,17 +38,13 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
-    dataBinding {
-        enabled = true
-    }
-
-
 }
 
 dependencies {
 
     implementation(project(":data:repository"))
     implementation(project(":model"))
+    implementation(project(":data:api"))
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.3.1'
@@ -130,4 +126,7 @@ dependencies {
 
     //whatif
     implementation "com.github.skydoves:whatif:1.1.0"
+
+    //coil
+    implementation("io.coil-kt:coil:1.4.0")
 }
