+++ app/src/main/res/layout/fragment_build_main.xml
@@ -1,7 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android">
-<?xml version="1.0" encoding="utf-8"?>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/build_main_frameLayout"
         android:layout_width="match_parent"
