+++ app/src/main/res/layout/item_build_main.xml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout>
     <data>
-
-
         <variable
             name="championData"
-            type="com.example.firstapp.model.Champion.Datum" />
+            type="com.example.firstapp.model.mychampion.Datum" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
@@ -16,15 +14,17 @@
     
     <de.hdodenhof.circleimageview.CircleImageView
         android:id="@+id/champion_image"
-        android:layout_width="50dp"
-        android:layout_height="50dp"
+        android:layout_width="70dp"
+        android:layout_height="70dp"
         android:src="@mipmap/ic_launcher"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:civ_border_width="2dp"
-        app:civ_border_color="#FFFFFF"/>
+        app:civ_border_color="#FFFFFF"
+        app:setImageView="@{championData.id}"
+        />
 
     <TextView
         android:id="@+id/champion_name"
