+++ build.gradle
@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
 
-    ext.hilt_version = "2.38.1"
+    ext.hilt_version = "2.39.1"
 
     repositories {
         google()
