+++ data/api/build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'com.android.library'
     id 'kotlin-android'
     id 'kotlin-kapt'
-    id 'kotlin-parcelize'
+
 }
 
 android {
@@ -54,5 +54,11 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 
+    
+    //sandwich
+    implementation "com.github.skydoves:sandwich:1.2.1"
+
+    //whatif
+    implementation "com.github.skydoves:whatif:1.1.0"
 
 }
