+++ data/repository/build.gradle
@@ -37,15 +37,19 @@ dependencies {
     implementation(project(":data:api"))
     implementation(project(":model"))
 
+
+    //noinspection GradleDependency
     implementation "com.google.dagger:hilt-android:$hilt_version"
-    implementation project(path: ':app')
+
     kapt "com.google.dagger:hilt-android-compiler:$hilt_version"
 
     implementation 'androidx.core:core-ktx:1.7.0'
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.2'
 
+    
     //sandwich
     implementation "com.github.skydoves:sandwich:1.2.1"
