+++ model/src/main/java/com/example/firstapp/model/mychampion/Datum.kt
@@ -1,4 +1,4 @@
-package com.example.firstapp.model.Champion
+package com.example.firstapp.model.mychampion
 
 import android.os.Parcelable
